rename voms_authz directory to authz directory
