+++ cvmfs/authz/voms_cred.h
